--- src/main/java/ru/spb/sspk/ssdmd/phonebook/PhoneBookApplication.java
@@ -1,13 +1,0 @@
-package ru.spb.sspk.ssdmd.phonebook;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PhoneBookApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PhoneBookApplication.class, args);
-	}
-
-}
