--- src/main/java/ru/spb/sspk/ssdmd/phonebook/model/dto/UserDto.java
@@ -1,23 +1,0 @@
-package ru.spb.sspk.ssdmd.phonebook.model.dto;
-
-public class UserDto {
-    private Long userId;
-
-    public UserDto() {
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    @Override
-    public String toString() {
-        return "UserDto{" +
-                ", userId=" + userId +
-                '}';
-    }
-}
