--- src/main/java/ru/spb/sspk/ssdmd/phonebook/repository/PersonRepository.java
@@ -1,22 +1,0 @@
-package ru.spb.sspk.ssdmd.phonebook.repository;
-
-import org.springframework.stereotype.Repository;
-import ru.spb.sspk.ssdmd.phonebook.model.entity.Person;
-
-import java.util.List;
-import java.util.Map;
-
-@Repository
-public interface PersonRepository {
-
-
-    Person save(Person person);
-
-    Person update(Person person);
-
-    List<Person> findAll();
-
-    List<Person> findByAll(Map<String, Object> paramMap);
-
-    void deleteById(Long id);
-}
