--- src/main/java/ru/spb/sspk/ssdmd/phonebook/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package ru.spb.sspk.ssdmd.phonebook.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.spb.sspk.ssdmd.phonebook.model.entity.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User save(User user);
-
-    User findByUserId(Long userId);
-
-}
