+++ src/main/java/ru/spb/sspk/ssdmd/phonebook_test/PhoneBookTestApplication.java
@@ -1,0 +1,13 @@
+package ru.spb.sspk.ssdmd.phonebook_test;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PhoneBookTestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PhoneBookTestApplication.class, args);
+	}
+
+}
