+++ src/main/java/ru/spb/sspk/ssdmd/phonebook_test/model/dto/UserDto.java
@@ -1,0 +1,23 @@
+package ru.spb.sspk.ssdmd.phonebook_test.model.dto;
+
+public class UserDto {
+    private Long userId;
+
+    public UserDto() {
+    }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    @Override
+    public String toString() {
+        return "UserDto{" +
+                ", userId=" + userId +
+                '}';
+    }
+}
