--- src/test/java/ru/spb/sspk/ssdmd/phonebook/PhoneBookTestApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.spb.sspk.ssdmd.phonebook;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PhoneBookTestApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
