+++ src/test/java/ru/spb/sspk/ssdmd/phonebook_test/PhoneBookTestApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.spb.sspk.ssdmd.phonebook_test;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PhoneBookTestApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
